pyproject.toml
[tool.poetry]
name = "mysql_backup"
version = "0.1.0"
description = "Rolling MySQL backups"
authors = ["Daniel Fels <mysqlbkp@dfels.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.15.0"
paramiko = "^2.7.2"
scp = "^0.13.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
